Main svc compilations

+++ main-server/src/main/java/ru/practicum/user/service/UserServiceImpl.java
@@ -19,7 +19,7 @@ import java.util.List;
 @Slf4j
 @Service
 @AllArgsConstructor
-@Transactional(readOnly = true)
+@Transactional
 public class UserServiceImpl implements UserService {
 
     private final UserRepository userRepository;
@@ -78,6 +78,9 @@ public class UserServiceImpl implements UserService {
     @Override
     @Transactional
     public void delete(Long id) {
+        if (!userRepository.existsById(id)) {
+            throw new NotFoundException("Не найден пользователь с id: " + id);
+        }
         log.info("Удаление пользователя с id: {}", id);
         userRepository.deleteById(id);
     }
